requirements.txt
langchain-openai
langchain
langchain-core
langchain-community
python-dotenv
openai
